requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
RPi.GPIO==0.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.16.2
